refactor(prices): rename fetcher-adapter & move much logic into application
