+++ prices/adapters/fetcher-adapter/src/exceptions/empty-yandex-data.exception.ts
@@ -1,0 +1,7 @@
+import { FetcherAdapterExceptionMessage } from './exceptions.constants.js'
+
+export class EmptyYandexDataError extends Error {
+  constructor() {
+    super(FetcherAdapterExceptionMessage.DOCUMENT_IS_EMPTY)
+  }
+}
