+++ prices/adapters/fetcher-adapter/src/exceptions/exceptions.constants.ts
@@ -1,0 +1,4 @@
+export enum FetcherAdapterExceptionMessage {
+  DOCUMENT_IS_EMPTY = 'Document is empty',
+  RESPONSE_IS_NULL = 'Respose is nullish',
+}
