+++ prices/adapters/fetcher-adapter/src/exceptions/index.ts
@@ -1,0 +1,2 @@
+export * from './null-response.exception.js'
+export * from './empty-yandex-data.exception.js'
