+++ prices/adapters/fetcher-adapter/src/exceptions/null-response.exception.ts
@@ -1,0 +1,7 @@
+import { FetcherAdapterExceptionMessage } from './exceptions.constants.js'
+
+export class NullResponseError extends Error {
+  constructor(url: string) {
+    super(`${FetcherAdapterExceptionMessage.RESPONSE_IS_NULL}, on ${url}`)
+  }
+}
