+++ prices/adapters/fetcher-adapter/src/services/index.ts
@@ -1,0 +1,1 @@
+export * from './fetch.service.js'
