--- prices/adapters/undici-adapter/src/constants/fetch-url.constants.ts
@@ -1,3 +1,0 @@
-export const API_URL = process.env.API_URL || 'https://api.aqsi.ru/pub/v2'
-export const GOODS_LIST_PATH = '/Goods/list'
-export const GOODS_CATEGORY_PATH = '/GoodsCategory/list'
