--- prices/adapters/undici-adapter/src/errors/empty-yandex-data.error.ts
@@ -1,5 +1,0 @@
-export class EmptyYandexDataError extends Error {
-  constructor() {
-    super(`yandex data is empty`)
-  }
-}
