--- prices/adapters/undici-adapter/src/errors/index.ts
@@ -1,2 +1,0 @@
-export * from './null-response.error.js'
-export * from './empty-yandex-data.error.js'
