--- prices/adapters/undici-adapter/src/errors/null-response.error.ts
@@ -1,5 +1,0 @@
-export class NullResponseError extends Error {
-  constructor(url: string) {
-    super(`Error on globals_aqsi-data-getter, response on ${url} is nullish`)
-  }
-}
