--- prices/adapters/undici-adapter/src/helpers/index.ts
@@ -1,1 +1,0 @@
-export * from './fetch-data.helper.js'
