--- prices/adapters/undici-adapter/src/interfaces/2gis-offers.interface.ts
@@ -1,22 +1,0 @@
-type TwoGisOffersItem = {
-  _attributes: {
-    id: string
-  }
-  name: {
-    _text: string
-  }
-  price: {
-    _text: string
-  }
-  currencyId: {
-    _text: string
-  }
-  categoryId: {
-    _text: string
-  }
-  description: {
-    _text: string
-  }
-}
-
-export type TwoGisOffersType = Array<TwoGisOffersItem>
