--- prices/adapters/undici-adapter/src/interfaces/categories.interface.ts
@@ -1,8 +1,0 @@
-type CategoriesItem = {
-  _attributes: {
-    id: string
-  }
-  _text: string
-}
-
-export type CategoriesType = Array<CategoriesItem>
