--- prices/adapters/undici-adapter/src/interfaces/goods-category-data-formatted.interface.ts
@@ -1,6 +1,0 @@
-type GoodsCategoryDataFormattedItem = {
-  id: string
-  name: string
-}
-
-export type GoodsCategoriesDataFormattedType = Array<GoodsCategoryDataFormattedItem>
