--- prices/adapters/undici-adapter/src/interfaces/goods-category-data.interface.ts
@@ -1,7 +1,0 @@
-type GoodsCategoryDataItem = {
-  id: string
-  name: string
-  children: Array<any>
-}
-
-export type GoodsCategoryDataType = Array<GoodsCategoryDataItem>
