--- prices/adapters/undici-adapter/src/interfaces/goods-data-formatted.interface.ts
@@ -1,2 +1,0 @@
-import type { RowType } from './goods-data.interface.js'
-export type GoodsDataFormattedType = Array<RowType>
