--- prices/adapters/undici-adapter/src/interfaces/goods-data.interface.ts
@@ -1,11 +1,0 @@
-export type RowType = {
-  id: string
-  group_id: string
-  name: string
-  price: string
-}
-
-export type GoodsDataType = {
-  pages: number
-  rows: Array<RowType>
-}
