--- prices/adapters/undici-adapter/src/interfaces/index.ts
@@ -1,8 +1,0 @@
-export * from './goods-data.interface.js'
-export * from './goods-category-data.interface.js'
-export * from './goods-data-formatted.interface.js'
-export * from './goods-category-data-formatted.interface.js'
-export * from './categories.interface.js'
-export * from './yandex-offers.interface.js'
-export * from './2gis-offers.interface.js'
-export * from './yml-data-type.interface.js'
