--- prices/adapters/undici-adapter/src/mappers/index.ts
@@ -1,1 +1,0 @@
-export * from './goods-fetch-promises.mapper.js'
