--- prices/adapters/undici-adapter/src/services/index.ts
@@ -1,3 +1,0 @@
-export * from './get-goods-data.service.js'
-export * from './get-goods-category-data.service.js'
-export * from './get-goods-page-data.service.js'
